requirements.txt
python-telegram-bot[job-queue]
websocket-client
requests
python-dotenv
telethon
